requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0

# FastAPI backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-frontmatter>=1.0.0
aiofiles>=23.0.0
